+++ common/src/main/java/tocraft/walkers/traits/impl/FearedTrait.java
@@ -3,14 +3,11 @@ package tocraft.walkers.traits.impl;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import org.jetbrains.annotations.NotNull;
-import tocraft.craftedcore.patched.CRegistries;
-import tocraft.craftedcore.patched.Identifier;
 import tocraft.walkers.Walkers;
 import tocraft.walkers.traits.ShapeTrait;
 
@@ -21,7 +18,6 @@ import java.util.function.Predicate;
 @SuppressWarnings("unused")
 public class FearedTrait<E extends LivingEntity> extends ShapeTrait<E> {
     public static final ResourceLocation ID = Walkers.id("feared");
-    @SuppressWarnings("unchecked")
     public static final MapCodec<FearedTrait<?>> CODEC = RecordCodecBuilder.mapCodec((instance) -> instance.group(
             Codec.list(ResourceLocation.CODEC).optionalFieldOf("fearful", new ArrayList<>()).forGetter(o -> o.fearfulTypes.stream().map(o1 -> Walkers.getEntityTypeRegistry().getKey(o1)).toList()),
             Codec.list(ResourceLocation.CODEC).optionalFieldOf("fearful_tags", new ArrayList<>()).forGetter(o -> o.fearfulTags.stream().map(TagKey::location).toList())
@@ -30,11 +26,11 @@ public class FearedTrait<E extends LivingEntity> extends ShapeTrait<E> {
         List<TagKey<EntityType<?>>> fearfulTags = new ArrayList<>();
         for (ResourceLocation resourceLocation : preyLocations) {
             if (Walkers.getEntityTypeRegistry().containsKey(resourceLocation)) {
-                fearfulTypes.add((EntityType<?>) Walkers.getEntityTypeRegistry().get(resourceLocation));
+                fearfulTypes.add(Walkers.getEntityTypeRegistry().get(resourceLocation));
             }
         }
         for (ResourceLocation preyTagLocation : preyTagLocations) {
-            fearfulTags.add((TagKey<EntityType<?>>) TagKey.create(Walkers.getEntityTypeRegistry().key(), preyTagLocation));
+            fearfulTags.add(TagKey.create(Walkers.getEntityTypeRegistry().key(), preyTagLocation));
         }
         return new FearedTrait<>(new ArrayList<>(), fearfulTypes, new ArrayList<>(), fearfulTags);
     })));
@@ -43,6 +39,7 @@ public class FearedTrait<E extends LivingEntity> extends ShapeTrait<E> {
     private final List<EntityType<?>> fearfulTypes;
     private final List<Class<? extends LivingEntity>> fearfulClasses;
     private final List<TagKey<EntityType<?>>> fearfulTags;
+    private final int priority;
 
     public static FearedTrait<?> ofFearfulType(EntityType<?>... fearful) {
         return new FearedTrait<>(new ArrayList<>(), List.of(fearful), new ArrayList<>(), new ArrayList<>());
@@ -63,10 +60,15 @@ public class FearedTrait<E extends LivingEntity> extends ShapeTrait<E> {
     }
 
     public FearedTrait(@NotNull List<Predicate<LivingEntity>> fearfulPredicates, @NotNull List<EntityType<?>> fearfulTypes, @NotNull List<Class<? extends LivingEntity>> fearfulClasses, @NotNull List<TagKey<EntityType<?>>> fearfulTags) {
+        this(fearfulPredicates, fearfulTypes, fearfulClasses, fearfulTags, 3);
+    }
+
+    public FearedTrait(@NotNull List<Predicate<LivingEntity>> fearfulPredicates, @NotNull List<EntityType<?>> fearfulTypes, @NotNull List<Class<? extends LivingEntity>> fearfulClasses, @NotNull List<TagKey<EntityType<?>>> fearfulTags, int priority) {
         this.fearfulPredicates = fearfulPredicates;
         this.fearfulTypes = fearfulTypes;
         this.fearfulClasses = fearfulClasses;
         this.fearfulTags = fearfulTags;
+        this.priority = priority;
     }
 
     public boolean isFeared(LivingEntity entity) {
@@ -81,6 +83,10 @@ public class FearedTrait<E extends LivingEntity> extends ShapeTrait<E> {
             if (fearfulPredicate.test(entity)) return true;
         }
         return false;
+    }
+
+    public int getPriority() {
+        return priority;
     }
 
     @Override