+++ CHANGELOG.md
@@ -2,7 +2,9 @@ walkers 5.1
 ================
 
 - fix blacklist commands stack
+- add FearedTrait & PreyTrait for Alex's Mobs
 - fix ConcurrentModificationException on AbilityRegistry
+- rework registration of PreyTrait & FearedTrait
 - fix crash when clicking on player
 
 walkers 5